cleanup C code

+++ RTL2832-2.2.2_kernel-3.0.0/demod_rtl2832.c
@@ -754,7 +754,7 @@ rtl2832_GetSignalStrength(
 	)
 {
 	unsigned long FsmStage;
-	int IfAgc;
+	long IfAgc;
