+++ RTL2832-2.2.2_kernel-3.0.0/nim_rtl2832_tda18272.c
@@ -369,7 +369,7 @@ rtl2832_tda18272_GetRfPowerLevelDbm(
 	DVBT_DEMOD_MODULE *pDemod;
 
 	unsigned long FsmStage;
-	int IfAgc;
+	long IfAgc;
 
 
 	// Get demod module.
